+++ spec/wavefront-sdk/dashboard_spec.rb
@@ -22,15 +22,14 @@ DASHBOARD_BODY = {
   ]
 }.freeze
 
-U_ACL_1 = { name: 'someone@example.com', id: 'someone@example.com' }
-U_ACL_2 = { name: 'other@elsewhere.com', id: 'other@elsewhere.com' }
+U_ACL_1 = { name: 'someone@example.com', id: 'someone@example.com' }.freeze
+U_ACL_2 = { name: 'other@elsewhere.com', id: 'other@elsewhere.com' }.freeze
 GRP_ACL = { name: 'example group',
-            id:   'f8dc0c14-91a0-4ca9-8a2a-7d47f4db4672' }
+            id:   'f8dc0c14-91a0-4ca9-8a2a-7d47f4db4672' }.freeze
 
 # Unit tests for dashboard class
 #
 class WavefrontDashboardTest < WavefrontTestBase
-=begin
   def test_list
     should_work(:list, 10, '?offset=10&limit=100')
   end
@@ -94,7 +93,6 @@ class WavefrontDashboardTest < WavefrontTestBase
                                           nil], :post, POST_HEADERS)
     should_be_invalid(:unfavorite)
   end
-=end
 
   def test_acls
     should_work(:acls, [%w[dash1 dash2]], 'acl?id=dash1&id=dash2')
@@ -103,12 +101,12 @@ class WavefrontDashboardTest < WavefrontTestBase
   def test_acl_add
     should_work(:acl_add, [DASHBOARD, [U_ACL_1, U_ACL_2], [GRP_ACL]],
                 'acl/add', :post, {}, acl_body(DASHBOARD,
-                                    [U_ACL_1, U_ACL_2],
-                                    [GRP_ACL]))
+                                               [U_ACL_1, U_ACL_2],
+                                               [GRP_ACL]))
 
     should_work(:acl_add, [DASHBOARD, [U_ACL_1, U_ACL_2]],
                 'acl/add', :post, {}, acl_body(DASHBOARD,
-                                    [U_ACL_1, U_ACL_2]))
+                                               [U_ACL_1, U_ACL_2]))
     assert_raises(ArgumentError) { wf.acl_add(DASHBOARD, U_ACL_1) }
     assert_raises(ArgumentError) do
       wf.acl_add(DASHBOARD, [U_ACL_1], GRP_ACL)
@@ -118,28 +116,28 @@ class WavefrontDashboardTest < WavefrontTestBase
   def test_acl_remove
     should_work(:acl_delete, [DASHBOARD, [U_ACL_1, U_ACL_2], [GRP_ACL]],
                 'acl/remove', :post, {}, acl_body(DASHBOARD,
-                                    [U_ACL_1, U_ACL_2],
-                                    [GRP_ACL]))
+                                                  [U_ACL_1, U_ACL_2],
+                                                  [GRP_ACL]))
 
     should_work(:acl_delete, [DASHBOARD, [U_ACL_1, U_ACL_2]],
                 'acl/remove', :post, {}, acl_body(DASHBOARD,
-                                    [U_ACL_1, U_ACL_2]))
+                                                  [U_ACL_1, U_ACL_2]))
     assert_raises(ArgumentError) { wf.acl_delete(DASHBOARD, U_ACL_1) }
   end
 
   def test_acl_set
     should_work(:acl_set, [DASHBOARD, [U_ACL_1, U_ACL_2], [GRP_ACL]],
                 'acl/set', :put, {}, acl_body(DASHBOARD,
-                                    [U_ACL_1, U_ACL_2],
-                                    [GRP_ACL]))
+                                              [U_ACL_1, U_ACL_2],
+                                              [GRP_ACL]))
 
     should_work(:acl_set, [DASHBOARD, [U_ACL_1, U_ACL_2]],
                 'acl/set', :put, {}, acl_body(DASHBOARD,
-                                    [U_ACL_1, U_ACL_2]))
+                                              [U_ACL_1, U_ACL_2]))
     assert_raises(ArgumentError) { wf.acl_set(DASHBOARD, U_ACL_1) }
   end
 
   def acl_body(id, view = [], modify = [])
-    { entityId: id, viewAcl: view, modifyAcl: modify }.to_json
+    [{ entityId: id, viewAcl: view, modifyAcl: modify }].to_json
   end
 end